graphviz: add a class to the generated SVGs
The generated SVGs now gain the `graphviz` class, which makes it easier
to customize their look via CSS. This is especially important when
adding support for dark mode, as there are other SVGs on the front page
that should not receive the same treatment.

Sadly, `node-html-parser` is not smart enough to handle
`element.classList.add(...)` followed by `element.toString()` correctly:
The output would _not_ have the added class. But there's always crude
string processing to save the day.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>