Use Font Awesome for homepage icons
There was no "Learn" or "Videos" icon, and I wanted the icons to match,
so we need new icons! There are GPL compatible according to their
website.

Integrated these icons as SVGs (just copied and pasted from their site)
to keep file sizes down, the main Font Awesome CSS file is pretty big.

This should make it easier to update the icons in the future. Switch
from using `float:` to `display: grid` to display the icons.